// apps/api/migrations/snapshots/c9a1bc7aa8b2407decf074636d7558fd1eb6fab28becd7c2fc3e6974e2b8f460/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Cart: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
    readonly CartItem: {
      readonly cartId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly productId: CodecTypes['pg/int4@1']['output'];
      readonly quantity: CodecTypes['pg/int4@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Category: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Order: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly deliveryAddress: CodecTypes['pg/text@1']['output'];
      readonly deliveryFee: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly paymentMethod: CodecTypes['pg/text@1']['output'] | null;
      readonly paymentStatus: CodecTypes['pg/text@1']['output'];
      readonly status: CodecTypes['pg/text@1']['output'];
      readonly subtotal: CodecTypes['pg/int4@1']['output'];
      readonly total: CodecTypes['pg/int4@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
    readonly OrderItem: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly orderId: CodecTypes['pg/int4@1']['output'];
      readonly productId: CodecTypes['pg/int4@1']['output'];
      readonly quantity: CodecTypes['pg/int4@1']['output'];
      readonly unitPrice: CodecTypes['pg/int4@1']['output'];
    };
    readonly Post: {
      readonly authorId: CodecTypes['pg/int4@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Product: {
      readonly categoryId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly isActive: CodecTypes['pg/bool@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly price: CodecTypes['pg/int4@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly stockQuantity: CodecTypes['pg/int4@1']['output'];
      readonly unit: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'] | null;
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Cart: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
    };
    readonly CartItem: {
      readonly cartId: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly productId: CodecTypes['pg/int4@1']['input'];
      readonly quantity: CodecTypes['pg/int4@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Category: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Order: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly deliveryAddress: CodecTypes['pg/text@1']['input'];
      readonly deliveryFee: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly paymentMethod: CodecTypes['pg/text@1']['input'] | null;
      readonly paymentStatus: CodecTypes['pg/text@1']['input'];
      readonly status: CodecTypes['pg/text@1']['input'];
      readonly subtotal: CodecTypes['pg/int4@1']['input'];
      readonly total: CodecTypes['pg/int4@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
    };
    readonly OrderItem: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly orderId: CodecTypes['pg/int4@1']['input'];
      readonly productId: CodecTypes['pg/int4@1']['input'];
      readonly quantity: CodecTypes['pg/int4@1']['input'];
      readonly unitPrice: CodecTypes['pg/int4@1']['input'];
    };
    readonly Post: {
      readonly authorId: CodecTypes['pg/int4@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Product: {
      readonly categoryId: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly isActive: CodecTypes['pg/bool@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly price: CodecTypes['pg/int4@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly stockQuantity: CodecTypes['pg/int4@1']['input'];
      readonly unit: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'] | null;
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly Cart: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
    readonly CartItem: {
      readonly cartId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly productId: CodecTypes['pg/int4@1']['output'];
      readonly quantity: CodecTypes['pg/int4@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Category: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Order: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly deliveryAddress: CodecTypes['pg/text@1']['output'];
      readonly deliveryFee: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly paymentMethod: CodecTypes['pg/text@1']['output'] | null;
      readonly paymentStatus: CodecTypes['pg/text@1']['output'];
      readonly status: CodecTypes['pg/text@1']['output'];
      readonly subtotal: CodecTypes['pg/int4@1']['output'];
      readonly total: CodecTypes['pg/int4@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
    };
    readonly OrderItem: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly orderId: CodecTypes['pg/int4@1']['output'];
      readonly productId: CodecTypes['pg/int4@1']['output'];
      readonly quantity: CodecTypes['pg/int4@1']['output'];
      readonly unitPrice: CodecTypes['pg/int4@1']['output'];
    };
    readonly Post: {
      readonly authorId: CodecTypes['pg/int4@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Product: {
      readonly categoryId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly isActive: CodecTypes['pg/bool@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly price: CodecTypes['pg/int4@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly stockQuantity: CodecTypes['pg/int4@1']['output'];
      readonly unit: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'] | null;
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;